requirements.txt
cryptography==2.1.4
PyYAML==3.12
